refactor: バリデーション関数をMathとToolからValidateに移動し、インポートパスを更新
