+++ package/main/src/Math/calculator/core.ts
@@ -2,7 +2,7 @@ import { addition } from "@/Math/addition";
 import { division } from "@/Math/division";
 import { multiplication } from "@/Math/multiplication";
 import { subtract } from "@/Math/subtract";
-import { isNumber } from "../isNumber";
+import { isNumber } from "@/Validate/isNumber";
 import { convertCurrency } from "./convertCurrency";
 
 export const calculatorCore = <T extends { [key: string]: string | number }>(
