+++ package/main/src/Math/index.ts
@@ -8,9 +8,6 @@ export * from "./factorial";
 export * from "./factorize";
 export * from "./gcd";
 export * from "./getDecimalLength";
-export * from "./isDouble";
-export * from "./isNumber";
-export * from "./isPrimeNumber";
 export * from "./isValueNaN";
 export * from "./lcm";
 export * from "./mathConverter";
