+++ package/main/src/Tool/index.ts
@@ -3,10 +3,4 @@ export * from "./createPipeline";
 export * from "./dayOfWeek";
 export * from "./deleteSpace";
 export * from "./hexToRgba";
-export * from "./isArray";
-export * from "./isBrowser";
-export * from "./isNode";
-export * from "./isNodeWebkit";
-export * from "./isNotEmpty";
-export * from "./isObject";
 export * from "./rgbaToHexA";
