+++ package/main/src/Validate/index.ts
@@ -1,0 +1,9 @@
+export * from "./isArray";
+export * from "./isBrowser";
+export * from "./isDouble";
+export * from "./isNode";
+export * from "./isNodeWebkit";
+export * from "./isNotEmpty";
+export * from "./isNumber";
+export * from "./isObject";
+export * from "./isPrimeNumber";
