+++ package/main/src/Validate/isArray.ts
