+++ package/main/src/Validate/isBrowser.ts
