+++ package/main/src/Validate/isDouble.ts
