+++ package/main/src/Validate/isNode.ts
