+++ package/main/src/Validate/isNodeWebkit.ts
