+++ package/main/src/Validate/isNotEmpty.ts
