+++ package/main/src/Validate/isNumber.ts
