+++ package/main/src/Validate/isObject.ts
