+++ package/main/src/Validate/isPrimeNumber.ts
